+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-### 2.6.0-beta
-* Ability to store a setting in different file/container/usersettings (simply pass file name in as last parameter)
-* Fixed issues with double parsing on Android
+### 2.6.0
+* Added ability to specify file name
+* iOS: SuiteName
+* Android: SharedPreferences name
+* UWP/WinRT Container Name
 
 
 ### 2.5.8